Change main page content and routing in routes.js
Modified the main page content in routes.js to display a list of users instead of a message input box. Updated the routing to handle '/users' and '/create-user' endpoints effectively. This was done to improve user interaction by listing all the users on the main page and enhancing data submission through the creation of a specific endpoint for user creation.